Split botlistener in multiple classes
